+++ gtest/src/lib.rs
@@ -427,7 +427,7 @@ mod mailbox;
 mod manager;
 mod program;
 mod system;
-mod block_info;
+mod blocks;
 
 pub use crate::log::{CoreLog, Log, RunResult};
 pub use codec;
